+++ Calculatrice_Et_Interface/NEG.java
@@ -2,7 +2,7 @@ class NEG extends EXPR_UNARY {
     NEG(EXPR v){
         this.value = v;
     }
-    int eval()
+    double eval()
     {
         return (-1)*(value.eval());
     }
