+++ Calculatrice_Et_Interface/SUB.java
@@ -4,7 +4,7 @@ class SUB extends EXPR_BINARY {
         this.right = r;
     }
 
-    int eval()
+    double eval()
     {
         return left.eval() - right.eval();
     }
